The following is fixed
- Fix Bluetooth Service to use Byte only
- Fix DataFile/Controller, DataSet/Controller, Meter/Controller to work with views
- Fixed list adaptors to work with data from database
- Link Views  with Contollers
- Store and retrive data form Database
- Fix issue when bluetooth is does not connect again to the same device
- Update Graph to show real data from DataFile
- Updated UI to fix fields name and create button

+++ SoundLevelApp/app/src/main/java/ca/concordia/teamc/soundlevelapp/DataSet.java
@@ -56,7 +56,7 @@ public class DataSet {
         return dateOfDownload;
     }
 
-    public void setDateOfDownload(int dateOfDownload) {
+    public void setDateOfDownload(long dateOfDownload) {
         this.dateOfDownload = dateOfDownload;
     }
 
@@ -64,7 +64,7 @@ public class DataSet {
         return dateStartRecord;
     }
 
-    public void setDateStartRecord(int dateStartRecord) {
+    public void setDateStartRecord(long dateStartRecord) {
         this.dateStartRecord = dateStartRecord;
     }
